+++ src/main/java/springboot/todoapp/repository/OrgRepository.java
@@ -3,5 +3,5 @@ package springboot.todoapp.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import springboot.todoapp.model.Org;
 
-public interface OrgRepository extends JpaRepository<Org, Integer> {
+public interface OrgRepository extends JpaRepository<Org, Long> {
 }
