+++ src/main/java/springboot/todoapp/repository/PurposeRepository.java
@@ -3,5 +3,5 @@ package springboot.todoapp.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import springboot.todoapp.model.Purpose;
 
-public interface PurposeRepository extends JpaRepository<Purpose, Integer> {
+public interface PurposeRepository extends JpaRepository<Purpose, Long> {
 }
