+++ src/main/java/springboot/todoapp/service/PurposeService.java
@@ -21,7 +21,7 @@ public class PurposeService {
     @Autowired
     private BoardRepository boardRepository;
 
-    public Org addPurpose(Integer boardId){
+    public Org addPurpose(Long boardId){
         Board board = boardRepository.findById(boardId).get();
         Purpose purpose = new Purpose();
         purpose.setTitle("untitled purpose");
@@ -31,13 +31,13 @@ public class PurposeService {
     }
 
     public Org updatePurpose(Purpose purpose){
-        Purpose purpose1 = purposeRepository.findById(purpose.getId().intValue()).get();
+        Purpose purpose1 = purposeRepository.findById(purpose.getId()).get();
         purpose1.setTitle(purpose.getTitle());
         purpose1.setContext(purpose.getContext());
         return purposeRepository.save(purpose1).getBoard().getOrg();
     }
 
-    public Org deletePurpose(Integer purposeId){
+    public Org deletePurpose(Long purposeId){
         System.out.println("PurposeService.deletePurpose");
         Purpose purpose = purposeRepository.findById(purposeId).get();
         Org org  = purpose.getBoard().getOrg();
